requirements.txt
streamlit==1.28.1
Pillow==10.0.1
pathlib
pyttsx3==2.90
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
openai==1.3.0
moviepy==1.0.3